[Temp] Demonstrate build with signing
The changes to .circleci/config.yml would, under this plan, be
made.  The Dev_Debug build would be certified and provisioned under
a new Development certificate and profile.

The creation of a Debug build and the use of my personal
certificate, provisioning profile, and bundle ID are all artifacts
of this demonstration and not intended for production.