CircleCI: Print Homebrew Info
Delete the prior CircleCI script, which is mostly valid but cannot run
on MacOS 10.15 Catalina.  The issue is that Catalina is stricter with
updates to system components such as Ruby.  The old script is no longer
allowed to install or update system Ruby gems.  The script functionality
will return in future commits to be merged at the same time as this
commit.

The first step of the new script will be to print Homebrew information
so that future issues can be diagnosed more quickly.

Testing: passes CircleCI