CircleCI: Install User Ruby
On MacOS 10.15 Catalina, we need to use a newer version of Ruby and gem
than the system version.  We install a user version of Ruby (and
implicitly gem), then add the Ruby and gem paths to .bash_profile.
CircleCI runs each command in a separate shell, so the settings in
.bash_profile will be available to the next command as desired.

Testing: CircleCI passes