Issue #220: Use median exchange rate
The app prior to this change displayed the exchange rate as the mean of
the queried exchanges' rates.  After this change, the median is used
instead.

Tested manually with even and odd numbers of exchanges.  With even
numbers, the average of the two elements closest to the middle are taken
as is traditional.