Issue #222: Button refreshes exchange rate
Add functionality such that the refresh button refreshes the exchange
rate.  The implementation here is to add an additional notification and
listen in the two places where the exchange rate is used: the overview
screen and the exchange rate screen.

Testing: watched each page refresh due to the exchange rate being
updated after render.