Major: Fix CLI build and unit test
Unforunately I was running into issues running the unit tests on the
simulator from the CLI tool, which is a prerequisite for getting the
build up on Circle CI.  As I fixed issues, I seem to have unravelled
multiple problems.  The best solution I could manage involves moving to
the Swift Package Manager for available dependencies and using the
latest version of each of those packages.

The largest functional change is the conversion from GRPCSwift to GRPC,
which is a breaking update.  I have monitored the behavior of the app
under the updated library and behavior is passable.  I am seeing
'errors' from the gRPC connections getting dropped and restarted by the
system; I suspect this is a matter of the calls surviving longer than
intended.  The likely reality is that these problems already existed but
the older library did not report them.

Under the circumstances, which is that said code needs significant
changes anyway, I am leaving the problem as-is.  I can't fix issues as
easily without the test framework and tests to check for regressions.